push dask stuff to separate utility
